move: component 폴더 하위로 common 폴더 생성 및 파일 정리
